Exports domain models and fixes namespaces
