--- LOS/LOS.Bussiness/Entities/IdentityModels/ApplicationRole.cs
@@ -1,8 +1,0 @@
-﻿using Microsoft.AspNet.Identity.EntityFramework;
-using System.ComponentModel.DataAnnotations.Schema;
-
-namespace LOS.Bussiness.Entities.IdentityModels
-{
-	[NotMapped]
-	public class ApplicationRole : IdentityRole { }
-}
