+++ LOS/LOS.Bussiness/LOS.Bussiness.csproj
@@ -102,17 +102,7 @@
     <Reference Include="System.Xml" />
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="Entities\File.cs" />
-    <Compile Include="Entities\IdentityModels\ApplicationRole.cs" />
-    <Compile Include="Entities\CartItem.cs" />
-    <Compile Include="Entities\Category.cs" />
-    <Compile Include="Entities\Comment.cs" />
     <Compile Include="ApplicationContext.cs" />
-    <Compile Include="Entities\IdentityModels\ApplicationUser.cs" />
-    <Compile Include="Entities\IdentityModels\GenericIdentityExtensions.cs" />
-    <Compile Include="Entities\Image.cs" />
-    <Compile Include="Entities\Product.cs" />
-    <Compile Include="Entities\RatingLog.cs" />
     <Compile Include="IApplicationContext.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="Services\CartServices\CartService.cs" />
@@ -134,7 +124,12 @@
       <SubType>Designer</SubType>
     </None>
   </ItemGroup>
-  <ItemGroup />
+  <ItemGroup>
+    <ProjectReference Include="..\LOS.Domain.Models\LOS.Domain.Models.csproj">
+      <Project>{1545B92C-6AC4-47C0-BF03-A99E1EA9C61F}</Project>
+      <Name>LOS.Domain.Models</Name>
+    </ProjectReference>
+  </ItemGroup>
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
   <!-- To modify your build process, add your task inside one of the targets below and uncomment it. 
        Other similar extension points exist, see Microsoft.Common.targets.
