+++ LOS/LOS.Bussiness/Services/CartServices/ICartService.cs
@@ -1,20 +1,17 @@
-﻿using LOS.Bussiness.Entities;
-using System;
+﻿using LOS.Domain.Models.Entities;
 using System.Collections.Generic;
-using System.Linq;
-using System.Text;
 using System.Threading.Tasks;
 
 namespace LOS.Bussiness.Services.CartServices
 {
     public interface ICartService
     {
-		Task<List<Product>> GetCartAsync(string userId);
+        Task<List<Product>> GetCartAsync(string userId);
 
-		Task AddAsync(string userId, int productId);
+        Task AddAsync(string userId, int productId);
 
-		Task RemoveAsync(string userId, int productId);
+        Task RemoveAsync(string userId, int productId);
 
-		Task ClearAsync(string userId);
-	}
+        Task ClearAsync(string userId);
+    }
 }
