+++ LOS/LOS.Bussiness/Services/CategoryServices/CategoryService.cs
@@ -1,4 +1,4 @@
-﻿using LOS.Bussiness.Entities;
+﻿using LOS.Domain.Models.Entities;
 using System.Collections.Generic;
 using System.Data.Entity;
 using System.Linq;
