+++ LOS/LOS.Bussiness/Services/ProductServices/IProductService.cs
@@ -1,4 +1,4 @@
-﻿using LOS.Bussiness.Entities;
+﻿using LOS.Domain.Models.Entities;
 using System.Collections.Generic;
 using System.Threading.Tasks;
 
@@ -16,34 +16,34 @@ namespace LOS.Bussiness.Services.ProductServices
 
         Task DeleteAsync(int productID);
 
-		Task<PagedList<Product>> GetListByCategoryIdAsync(int pageNumber, int pageSize, int categoryID, string keyword= "", string sort = "");
+        Task<PagedList<Product>> GetListByCategoryIdAsync(int pageNumber, int pageSize, int categoryID, string keyword = "", string sort = "");
 
-		Task<PagedList<Product>> GetListByParentCategoryIdAsync(int pageNumber, int pageSize, int categoryID, string keyword = "", string sort = "");
+        Task<PagedList<Product>> GetListByParentCategoryIdAsync(int pageNumber, int pageSize, int categoryID, string keyword = "", string sort = "");
 
-		Task<List<Product>> GetAllAsync();
+        Task<List<Product>> GetAllAsync();
 
-		Task<List<Image>> GetImagesAsync(int productId);
+        Task<List<Image>> GetImagesAsync(int productId);
 
-		Task RateAsync(int productId, int rating, string userId);
+        Task RateAsync(int productId, int rating, string userId);
 
-		Task<decimal> GetRatingAsync(int productId);
+        Task<decimal> GetRatingAsync(int productId);
 
-		Task<int> GetVoteCountAsync(int productId);
+        Task<int> GetVoteCountAsync(int productId);
 
-		Task<bool> CheckIfUserHasVotedAsync(string userId, int productId );
+        Task<bool> CheckIfUserHasVotedAsync(string userId, int productId);
 
-		Task UpdateQuantityAsync(int productId);
+        Task UpdateQuantityAsync(int productId);
 
-		Task<List<Product>> GetLatestAsync(int productCount);
+        Task<List<Product>> GetLatestAsync(int productCount);
 
-		Task AddImageAsync(Image image);
+        Task AddImageAsync(Image image);
 
-		Task UploadFileAsync(Entities.File file);
+        Task UploadFileAsync(File file);
 
-		Task<bool> CheckIfImageExistsAsync(Image image);
+        Task<bool> CheckIfImageExistsAsync(Image image);
 
-		Task<Entities.File> GetFileByNameAsync(string fileName);
+        Task<File> GetFileByNameAsync(string fileName);
 
-		Task<List<Product>> GetRatedProductsForUserAsync(string userId);
-	}
+        Task<List<Product>> GetRatedProductsForUserAsync(string userId);
+    }
 }
