+++ LOS/LOS.Bussiness/Services/UserServices/IUserService.cs
@@ -1,15 +1,7 @@
-﻿using LOS.Bussiness.Entities.IdentityModels;
-using Microsoft.AspNet.Identity;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace LOS.Bussiness.Services.UserServices
+﻿namespace LOS.Bussiness.Services.UserServices
 {
-	public interface IUserService
-	{
+    public interface IUserService
+    {
 
-	}
+    }
 }
