+++ LOS/LOS.Bussiness/Services/UserServices/UserService.cs
@@ -1,15 +1,15 @@
-﻿using Microsoft.AspNet.Identity;
-using LOS.Bussiness.Entities.IdentityModels;
+﻿using LOS.Domain.Models.Entities.IdentityModels;
+using Microsoft.AspNet.Identity;
 
 namespace LOS.Bussiness.Services.UserServices
 {
-	public class UserService : UserManager<ApplicationUser>, IUserService
-	{
-		private IUserStore<ApplicationUser> store;
+    public class UserService : UserManager<ApplicationUser>, IUserService
+    {
+        private IUserStore<ApplicationUser> store;
 
-		public UserService(IUserStore<ApplicationUser> store) : base(store)
-		{
-			this.store = store;
-		}
-	}
+        public UserService(IUserStore<ApplicationUser> store) : base(store)
+        {
+            this.store = store;
+        }
+    }
 }
