+++ LOS/LOS.Domain.Models/Entities/Category.cs
@@ -1,15 +1,9 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace LOS.Bussiness.Entities
+﻿namespace LOS.Domain.Models.Entities
 {
     public class Category
     {
         public int CategoryID { get; set; }
         public int? ParentCategoryID { get; set; }
         public string Name { get; set; }
-	}
+    }
 }
