+++ LOS/LOS.Domain.Models/Entities/Comment.cs
@@ -1,10 +1,6 @@
 ﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 
-namespace LOS.Bussiness.Entities
+namespace LOS.Domain.Models.Entities
 {
     public class Comment
     {
