+++ LOS/LOS.Domain.Models/Entities/IdentityModels/ApplicationRole.cs
@@ -1,0 +1,8 @@
+﻿using Microsoft.AspNet.Identity.EntityFramework;
+using System.ComponentModel.DataAnnotations.Schema;
+
+namespace LOS.Domain.Models.Entities.IdentityModels
+{
+    [NotMapped]
+    public class ApplicationRole : IdentityRole { }
+}
