+++ LOS/LOS/Controllers/ErrorController.cs
@@ -1,53 +1,50 @@
-﻿using LOS.Bussiness.Entities;
-using LOS.Bussiness.Services.CategoryServices;
+﻿using LOS.Bussiness.Services.CategoryServices;
+using LOS.Domain.Models.Entities;
 using LOS.Models;
-using System;
 using System.Collections.Generic;
-using System.Linq;
 using System.Net;
 using System.Threading.Tasks;
-using System.Web;
 using System.Web.Mvc;
 
 namespace LOS.Controllers
 {
-	[AllowAnonymous]
+    [AllowAnonymous]
     public class ErrorController : Controller
     {
-		private readonly ICategoryService categoryService;
+        private readonly ICategoryService categoryService;
 
-		public ErrorController(ICategoryService categoryService)
-		{
-			this.categoryService = categoryService;
-		}
+        public ErrorController(ICategoryService categoryService)
+        {
+            this.categoryService = categoryService;
+        }
 
-		public ActionResult NotFound()
-		{
-			Response.TrySkipIisCustomErrors = true;
-			Response.StatusCode = (int)HttpStatusCode.NotFound;
-			return View("NotFound");
-		}
+        public ActionResult NotFound()
+        {
+            Response.TrySkipIisCustomErrors = true;
+            Response.StatusCode = (int)HttpStatusCode.NotFound;
+            return View("NotFound");
+        }
 
-		public ActionResult InternalServerError()
-		{
-			Response.TrySkipIisCustomErrors = true;
-			Response.StatusCode = (int)HttpStatusCode.NotFound;
-			return View("InternalServerError");
-		}
+        public ActionResult InternalServerError()
+        {
+            Response.TrySkipIisCustomErrors = true;
+            Response.StatusCode = (int)HttpStatusCode.NotFound;
+            return View("InternalServerError");
+        }
 
-		public PartialViewResult _Navbar()
-		{
-			List<Category> categories = Task.Run(async () =>
-			{
-				categories = await categoryService.GetAllAsync(); return categories;
-			}).Result;
+        public PartialViewResult _Navbar()
+        {
+            List<Category> categories = Task.Run(async () =>
+            {
+                categories = await categoryService.GetAllAsync(); return categories;
+            }).Result;
 
-			NavbarModel model = new NavbarModel
-			{
-				Categories = categories
-			};
+            NavbarModel model = new NavbarModel
+            {
+                Categories = categories
+            };
 
-			return PartialView("_Navbar", model);
-		}
-	}
+            return PartialView("_Navbar", model);
+        }
+    }
 }
