+++ LOS/LOS/Models/LatestProductsModel.cs
@@ -1,14 +1,11 @@
-﻿using LOS.Bussiness.Entities;
-using System;
+﻿using LOS.Domain.Models.Entities;
 using System.Collections.Generic;
-using System.Linq;
-using System.Web;
 
 namespace LOS.Models
 {
-	public class LatestProductsModel
-	{
-		public List<Product> Products { get; set; }
-		public List<Image> Images { get; set; }
-	}
+    public class LatestProductsModel
+    {
+        public List<Product> Products { get; set; }
+        public List<Image> Images { get; set; }
+    }
 }
